+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - fix some cases of terminal let in a bad state on errors
 - some changes to the fish shell function and its installation (PR #128)
 - **breaking change:** commands given with `--cmd` must be separated (default separator is `;`)
+- consider path `$ZDOTDIR/.zshrc` for zsh shell function sourcing (fix #90)
 
 <a name="v0.11.9"></a>
 ### v0.11.9 - 2020-01-15